requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.73.0
h11==0.13.0
idna==3.3
pydantic==1.9.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.0
uvicorn==0.17.4
psycopg2-binary==2.9.3